Mobile Client: Pipeline providers: calls the pai on change + collection is now private
