+++ mobile/lib/src/widgets/aeris_page.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:aeris/src/widgets/background/animated_background.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 /// Application base page, holds scaffold and background
 class AerisPage extends StatelessWidget {
